requirements.txt
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
pyTelegramBotApi>=4.2.0
jira>=3.1.1
funcy>=1.16
python-dotenv>=0.19.2